Visual Studio 2012 project files for those wanting to try it
This was NOT just a straightforward conversion.  It took days to
figure out why projects kept building over and over again, so
committing is the best way for people to kick the tires.

These are completely separate from the VS2008 project files, so
they do not affect current build process.