Meter display changes as suggested by Peter, Steve, and Bill
Adds a transition from green to yellow starting at -12 db
or (0.50 for linear) and ending at -6 (or .75).  A transition
to red then occors to the end of the meter.

Meter preferences have been added and most of the context
menu items have been moved there.

In addition, you may now click the playback meter to reset
the peak level indicator instead of having to click the
record meter.