Modules renamed as 'mods' for user.
We previously had both 'modules' and 'plug-ins' and the files are called things like mod-script-pipe.dll.  Three different things.  From a user's point of view it is easier to think of them as modifications to the basic program, like mods in Minecraft.  Hence the name change.  The search path now has 'mods' in it rather than 'modules'.  Didn't see a need to rename the 'Modules' source code as programmers know what modules are and the mods are modules too.