Fix runtime problem with wxWidgets 3.0: Dialog size assert. [Patch 0021] From martin@steghoefer.eu (adjusted by James).
Due to changes in sizing of dialogs with wx3, an assert about the size of the preferences dialog (was max 800x600) fails, which causes an error dialog to pop up. We've recently decided that screens have got bigger, and increased the default size of the Audacity window.  Rather than remove this assert, we've upped its limits so that we can go further before we hit the limit.  Other code to try to keep the dialog size near to 800x600 has not been changed.